refactor: move authStore to subfolder
