+++ src/context/authStore/authStore.tsx
@@ -6,9 +6,9 @@ import React, {
   useCallback,
 } from "react";
 import AsyncStorage from "@react-native-async-storage/async-storage";
-import { usePrevious } from "../hooks/usePrevious";
-import { AuthCredentials } from "../types";
-import { Sentry } from "../utils/errorTracking";
+import { usePrevious } from "../../hooks/usePrevious";
+import { AuthCredentials } from "../../types";
+import { Sentry } from "../../utils/errorTracking";
 import {
   readFromStoreInBuckets,
   saveToStoreInBuckets,
